DigitalOcean: Ignore terraform.tfvars
Using terraform.tfvars locally to define the value of a DigitalOcean API
token, it should not be checked in to source control. Therefore, we are
ignoring the `terraform.tfvars` file in order to just run `terraform
apply` automatically without providing a `-var-file` flag.

Also added an example file for user pulling the repository to replace
the variables' values required.